Categorize Scaffolds as buildings for the deconstruction tool
There have been false error reports that the deconstruction tool does
not work on scaffolds. This is because scaffolds are not categorized as
buildings, so players who filter to only deconstruct buildings are
surprised when scaffolds are not also deconstructed.

This commit patches the GetFilterLayerFromObjectLayer method to
categorize scaffolds as buildings.